swap up test formats for new config dict and mixed varitey to run of dbt parse and inspecting the manifest
